tensorflow_version/requirements_tf.txt
tensorflow>=2.10.0
pillow>=8.0.0
numpy>=1.19.0
matplotlib>=3.3.0 